requirements.txt
asgiref==3.5.0
Django==4.0
python-dotenv==0.19.2
sqlparse==0.4.2
python-telegram-bot==13.8.1
psycopg2-binary==2.9.1
